poly2vd: correct malfunctioning isEdgeDefinedByDummyPoint()
It used to produce false positives, because I erroneously
expected that it suffices to check site ids.

But instances of particular types of defining sites (points,
segments and arcs) are stored in separate vectors (defined
in Vroni's vddata.cc) -- pnts, segs and arcs. To detect
whether the defining site is a dummy point, one must check
the site type (dummy point has to be a point) and site id
(the first pair and the last pair of elements in the pnts
vector are the dummy points, according to the Vroni's
documentation)

Please refer to page four of the "VD Data Structure" paper
in Vroni's READMEs/README_VD_structure.pdf