poly2vd: introduce EdgeMeta and frontierGoalNode
During BFS, store "pointers" to previous nodes.

When doing the Breadth First Search, look for whether the edge is frontier based or not. When the
first frontier based edge is met on particular path leading from the root node, edges on that path
are published as ROS's Markers in the VdPublisher.appendPath() method. Nodes on this path are
intended to be examined as critical node candidates. During this process, expanded WMAT edges are
marked using EdgeStatus field of EdgeMeta records as either EXPLORED or FRONTIER_BASED.

The VdPublisher.appendPath() method also marks edges with nodeIds of the first reached frontier
based node connected to the root node by (the shortest) path that contains the particular edge.