Evaluate the critical node criteria during the backtracking phase
When exploring nodes on a path (during the "backtracking" phase),
evaluate the critical node criteria and publish critical node Markers
for Rviz. The critical node criteria evaluation is implemented in the
examineNode() function.

Fuzziness in clearance radii extremality criterion:

One of the original criteria for a critical node (in the
Burgard-Stachniss article) is that its clearance radius has to be a
local minimum compared to clearance radii of its neighbours.  We,
however, observed, that such a condition does no good, because often,
"good" critical node candidates are situated in narrow passages such
that their direct neighbours have exactly the same clearance radius as
the "good candidates" and a rigorous local minimum criterion would
reject them. We therefore employed a "fuzzy" clearance radius minimality
criterion implemented in the testNForBeingFuzzyMinimum() function and
described in the following paragraph.

When querying whether critical node candidate's radius is local
minimum, we query also its further neighbours in case when clearance
radii of the proximate ones are almost equal.

Other changes:

- mark some functions for removal (add FIXME comments): some of them are
  not used and isDeg2Node(int) is not correct, I suppose
- update TODO
- change colors (being published to Rviz) to more vivid ones